env: Add support for shell conditional syntax
Bash has support for some conditionals that allow to append text
to environment variables in a more clever way than just blindly
appending it. This is a must, for example, when appending paths
to PATH or LD_LIBRARY_PATH, because if the original variables
are empty, the new paths must be added "as-is", but if the
original variables have a value, a colon is required between
the old value and the new one.

This patch allows to append extra paths to the current defined
ones directly in the 'environment' section of the snapcraft.yaml
file.

The new supported syntaxes are:

  ${VARIABLE:-XXXXXX}

  if VARIABLE is defined, this will return the value of that
  variable; but if it is undefined, it will return the expansion
  of XXXXX (thus, if there are variables there, they will be
  recursively expanded).

  ${VARIABLE:+YYYYY}

  if VARIABLE is undefined, it will return an empty string;
  but if it is defined, it will return the expansion of YYYYY.

Example:

  ${PATH:+$PATH:}/usr/local/bin

  If PATH is undefined or empty, it will return an empty string,
  to with '/usr/local/bin' will be appended.
  But if PATH contains, for example, /usr/bin:/bin, it will
  return that plus a colon at the end, to which '/usr/local/bin'
  will be appended, thus resulting in the expected value

    /usr/bin:/bin:/usr/local/bin