overlord/ifacestate: emit snapd.apparmor warning only when using apparmor backend
The warning about snapd.apparmor not being enable should only be emitted
when apparmor backend is present, which happens when snapd has detected
a usable support for apparmor in the host system. A mere fact of having
the service packaged in the system is not sufficient to show the warning
to the user.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>