many: remove the gpio-chardev-interface experimental flag (#15449)
* many: remove the gpio-chardev-interface experimental flag

GPIO chardev support is hidden until kernel configfs gpio-aggregator
interface makes it to the 24.04 kernel AND the snap-gpio-helper is
updated to use the new configfs interface.

In the meantime remove the feature flag and mark the gpio-chardev
interface as unsupported to avoid snaps from using the workaround
implementation that utilized the gpio-aggregator sysfs kernel
interface so that it can be completely removed after configfs
support lands without having to worry about backwards compatibility.

* fixup! tests: check that gpio-chardev support is missing

* fixup! tests: comment out unreachable code in tests/core/interfaces-gpio-chardev

---------

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>