packaging, data/systemd: enable snapd.apparmor on Fedora (#15434)
* data/selinux: leave a TODO for snapd-apparmor

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* packaging/fedora: ship snapd-apparmor and related systemd service

Ship snapd-apparmor and the related systemd service.

This enables Fedora container to load the relevant apparmor profile when
running on an Ubuntu or other AppArmor enabled host.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* data/systemd/snapd: source snapd environment file

The snapd environment file can globally control reexec, make sure to
source it to get consistent behavior on distros which can explicitly
opt-in for reexec.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>