overlord/fdestate: give hint of encryption to read the PCR handle
When trying to read the potential PCR handle we try to read the key,
if it is not TPM, we return 0. Unfortunately, it is difficult to
discriminate hook v1 format keys (2.49, 2.50) from legacy tpm keys and
we need a hint. So we pass that hint from the sealed keys method (if
we find it, so we do not add too much dependency on it).