i/p/requestprompts: handle expired notification when replying (#15463)
* i/p/requestprompts: handle expired notification when replying

If a message notification has expired in the kernel, than an attempt to
reply to the corresponding request will result in ENOENT from the
kernel.

Previously, any error when sending the reply would result in aborting
the reply and surfacing the error to the caller.

This commit instead handles ENOENT differently, recording a debug log
and otherwise carrying on without returning an error. That means that if
a prompt reply receives an ENOENT for one of the listener requests
associated with that prompt, the other requests will still be sent
responses.

Additionally, if a reply returns ENOENT, then any rule which would be
created as a result of the successful reply will still be created.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestprompts: include request details in log of reply ENOENT

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>