cmd/snap-confine: do not drop capabilities if the original user is already root (#15441)
* cmd/snap-confine: do not drop permitted capabilities too early for root user

Skip dropping capabilities permitted set too early, if the original user
is already root. This preserves the original behavior we had before
introducing non-setuid support, where it was possible for the root user
to have its current working directory to be restored properly rigt
before we exec into the snap.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/main/classic-confinement-perms: add a test for classic confinement

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>