s/a/n/listener: improve test flakiness (#15476)
* s/a/n/listener: improve test flakiness

Before checking for a logger.Notice, explicitly wait for the callback
which records that log to finish.

Additionally, when the timeout callback is expected to not fire, throw a
fatal error if it does fire.

Lastly, increase timeouts for tests which expect something to happen
quickly but throw an error if it does not occur after that timeout.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/n/listener: avoid timer initialization race when no pending requests

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>