cmd/snap-confine: suport for host policy for limiting users able to run snaps
Some distributions have opinions on who can run snaps and would prefer
to have a way to limit execution of snaps only to members of a specific
group.

The patch allows support for setting the group ownership of snap-confine
as an indicator that only members of said group should be able to run
snaps.

In theory the same can be limited by setting root:<group> ownership and
then 0750 flags, or ACLs, but this only works for the local case.
However, this is not enough to ensure a consistent policy if reexec is
enabled. In such scenario /usr/bin/snap reexecs into
/snap/snapd/current/usr/bin/snap and proceeds to call snap-confine from
the snapd snap. The snapd snap is built as part of the upstream releases
and is not aware of any local distro policies expressed through group
ownership or ACLs, unless this is explicitly accounted for in the code,
which is done in this patch.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>