Allow to easily autocomplete internal apps
When there are several internal apps (snap_name.app_name), the
previous code required to autocomplete first the snap name, then
delete one character, add a dot, and only then would show all
the internal apps.

Now, when autocompleting, will only show snap names until there
is only one option to autocomplete. In that case, it will allow
to autocomplete with the snap name, or the internal apps. Of
course, if there is no default app (an app with the same name
than the snap), only the internal apps will be shown for auto-
completion.