secboot/secboot_hooks.go: do not require primary key on V1 format (#15469)
* secboot/secboot_hooks.go: do not require primary key on V1 format

V1 hook format can be only found when we read each key file we try to
reseal. There is no reseal in V1 hook format. And there is no primary
keys. So we should not fail reading primary keys if only V1 keys were
found.

* overlord/fdestate: give hint of encryption to read the PCR handle

When trying to read the potential PCR handle we try to read the key,
if it is not TPM, we return 0. Unfortunately, it is difficult to
discriminate hook v1 format keys (2.49, 2.50) from legacy tpm keys and
we need a hint. So we pass that hint from the sealed keys method (if
we find it, so we do not add too much dependency on it).