i/builtin: allow checkbox-support to run other snaps
The checkbox-support interface gained an explicit profile transition
for /usr/lib/snapd/snap-confine, which when using core18+, contains the
updated snap toolchain from the host or from snapd snap.

The intent of the profile was to be effectively as unconfined as
possible, but this still requires explicit permission for profile
transitions.

Add an integration test to make sure this really works.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>