system-packages-doc: Fix access to /usr/local/share/doc (#15499)
With commit "8e8eb2ddce interfaces: grant access to /usr/local/share/doc",
access was granted to /usr/local/share/doc. However, the mount target does
not (or ceased to) exist in the base snap, so the snaps connecting this
interface did not yet have access to that directory.

LP: #1955325