i/p/requestprompts: restore prompts after snapd restart (#15224)
* i/p/requestprompts: restore prompts after snapd restart

Record a map from request ID to prompt ID so that when snapd restarts
and requests are re-received from the kernel, prompts can be created
which have the same IDs as those which were originally associated with
the requests.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestprompts: add function to prune request map when ready occurs

When the listener is ready, any requests in the ID map which have not
yet been re-received from the kernel can be discarded from the map,
since readying indicates that there are no more pending requests waiting
to be re-received.

In doing so, the requestprompts backend should also record a notice for
any prompt IDs in the map for which no mapped requests have been
re-received, and thus the corresponding prompt has not been re-created.

If snapd restarts before the ready is handled, then we want to preserve
all IDs in the map, so we can handle all requests which may be received
after the restart.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/i/apparmorprompting: ready prompt backend after listener readies

When the listener signals that all pending requests have been
re-received, and the manager has finished handling all those requests,
the manager should tell the prompts backend that it is ready. This
allows it to prune any unreceived requests from the ID map.

The manager should wait for the prompts backend to finish handling the
readying before it closes the manager-level ready channel, so that the
prompts backend is consistent before any potential API requests can hit
it.

The manager should only tell the prompts backend to ready if the
listener has signalled readiness, *not* if the manager is closing before
this occurs. This is because snapd is restarting quickly before all
requests have been re-received, and we want the opportunity to retry
re-receiving and handling all pending requests after snapd finishes
restarting.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestprompts: deduplicate listener requests associated with prompt

Ensure that if a request is received multiple times (such as because
`APPARMOR_NOTIF_RESEND` is called again), a response with that request
ID is not sent multiple times, and the request is not associated
multiple times with the prompt.

This is important to avoid the kernel returning an error to repeated
replies against the same request ID, and so that we don't potentially
try to access and delete the same request ID from the ID map multiple
times.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestprompts: refactor common checks when handling new request

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestprompts: refactor handling of existing prompt or ID mapping

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestprompts: replace bitflags with struct for existing prompt result

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestprompts: adjust findExistingPrompt result and consistency

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/p/requestprompts: refactor request handling in prompts and adjust saved mapping filename

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>