cmd/snap: Set Kerberos cache to hostfs path.
SSSD creates the user's Kerberos ticket in /tmp and sets
the KRB5CCNAME to point to that file.

Snaps have a private /tmp so they can't reach that file.

So prepend /var/lib/snapd/hostfs/ to KRB5CCNAME when running the snap,
which will still require a interface (kerberos-tickets) to read the
tickets from that path.