progress: fix progress does not work with CJK (#15230)
* progress: fix progress does not work with CJK

The display width of a rune may have 0, 1 or 2. Especially for CJK characters,
nearly all CJK characters are 2 wide.

The code before this patch assumes the width of all runes is 1, it causes the
width of a text which have CJK characters has been wrongly calculated. Thus,
the progress status text may overflow into next line on the terminal.

This patch fix this issue by using runewidth to detect the length of each rune.

Fixes: https://bugs.launchpad.net/snapd/+bug/1952500

Signed-off-by: Zhaoxuan Zhai <zhai.zhaoxuan@canonical.com>

* progress: fix typo in comment

Signed-off-by: Zhaoxuan Zhai <zhai.zhaoxuan@canonical.com>

* progress: avoid rpercent and rspeed been calculated using runeWidth

Signed-off-by: Zhaoxuan Zhai <zhai.zhaoxuan@canonical.com>

* packaging: add github.com/mattn/go-runewidth and github.com/rivo/uniseg

The second dependency is indriect.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* packaging: sort dependencies

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* progress: cleanup redundant code and add comments to …

Signed-off-by: Zhai Zhaoxuan <zhai.zhaoxuan@canonical.com>

* progress: fix msg is not cut correctly in Notify()

Signed-off-by: Zhaoxuan Zhai <zhai.zhaoxuan@canonical.com>

* go: downgrade go-runewidth to 0.0.15

This makes the dependency installable on Fedora 40

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: remove fedora 40 from CI

Fedora 40 has reached end-of-life status on 13th of May 2025
as listed on https://docs.fedoraproject.org/en-US/releases/eol/

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

---------

Signed-off-by: Zhaoxuan Zhai <zhai.zhaoxuan@canonical.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
Signed-off-by: Zhai Zhaoxuan <zhai.zhaoxuan@canonical.com>
Co-authored-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>