many: add notice manager and notice backends outside of state
To avoid the state lock being required whenever a notice is added or
notices are requested, add a notice manager in the overlord which
mediates requests to multiple notice backends. Define a notice backend
interface which can be implemented by entities which provide notices.
The state remains a notice backend, and there are additional backends
defined for prompts and rules.

The daemon can ask the notice manager for notices without taking state
lock, and the manager then only requests notices from the backends which
are registered to provide notices of the requested type. Thus, if only
prompt or rule notices are requested, the manager can ask only the
prompts and rules backends, avoiding the state lock. Similarly, those
backends can create notices in whichever manner they please, without
needing to go through the state at all (except to get a unique and
monotonically increasing notice timestamp, which does not require state
lock).

Notices must still have unique and monotonically increasing timestamps,
and since notices can be created outside of snapd state, getting a
unique timestamp cannot require state lock. As such, there is now a
dedicated mutex in the state protecting the last timestamp, and a method
which the state and other notice backends may use to get a timestamp.
Since the state is no longer the only notice provider, the last notice
timestamp may be from a notice outside of the snapd state which is
marshalled to disk. Because getting a unique notice timestamp does not
cause state to be written to disk, the last notice timestamp is no
longer included in the state marshalled to disk, to avoid inconsistency.

Notice backends are registered to the notice manager during startup with
notice types and unique namespaces, which are prefixes to notice IDs
generated by those backends. The snapd state is the only backend which
is registered with an empty namespace. In this way, each backend has
unique notice IDs, they do not have to go through state to get an ID,
and the notice manager knows which backend is capable of generating a
notice with a given ID and/or type.

For the prompt and rule notice backends, prompts/rules only have
sufficient metadata to derive the user, key, and ID of a corresponding
notice, not things like the number of occurrences, first occurred/last
repeated timestamps (these timestamps have different semantics than the
prompt/rule timestamps), etc. Additionally, notices must outlive
prompts/rules, since they are used to notify clients after a prompt has
been resolved or a rule deleted. Thus, rather than inject additional
notice metadata into prompts/rules, the notices are stored separately in
memory and on disk.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>