many: make confdb API errors consistent with options
This makes the confdb API errors more consistent with the options API.
Instead of a 404 for various "not found" errors, we return a 400 with an
appropriate error kind. This uses the same error kind that options uses
for the "no data" scenario and introduces a couple of others (view not
found, no matching rule, etc).

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>