overlord/devicestate,tests: set revision of seed components to x1
Make sure that when we read a seed, the local components get a x1 revision. Not
doing this was causing strange behaviors on installation, like having extra
components in /var/lib/snapd/snaps/ with revision "unset", and having those
components mounted too.

A check is spread tests is added for this.