many: set x1 as default rev for unasserted containers
Set revision to x1 (-1 internally) for unasserted snaps and components when
reading seeds and preseeding assertions. This is consistent with the version
that such containers get on installation, as these are installed for the first
version from the seed. Remove previous partial fix in
loadAndMountSystemLabelSnaps().

Not doing this was causing strange behaviors on installation, like having extra
components in /var/lib/snapd/snaps/ with revision "unset", and having those
components mounted too.

A check for this in spread tests is added.