packaging/pack-source: exclude .git directory
Exclude .git directory, not just its contents. This works around the
problem, where go build notices the presence of .git directory and
attempts to incorporate VCS information into the build, but since the
directory is empty, the attempt fails. One can pass `-buildvcs=false` to
disable this, but it needs to be included in the appropriate packaging
files when consuming tarballs generated by pack-source.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>