osutil/vfs: add behavior integration test
This test does not check the implementation but instead aids in review
of the Go logic, by providing executable examples observing specific
properties of the Linux kernel.

The tests post a question and then observe the answer against Debian 10,
running Linux 4.19, and Debian Sid, running at the time of writing this
message, Linux 6.12, trying to get the widest sensible span of time to
show that kernel behavior is indeed stable.

The test environment is very quick to start, allowing developers to
perform interactive exploration of the behavior.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>