many: add support for private /var/tmp
Historically snaps could only access /tmp in a specially-created
subdirectory of the real /tmp present on the host. This was done because
nearly every application needs to use /tmp in one way or another.

No consideration was originally made for /var/tmp and several cases that
came across time were dealt with by patching the application to use /tmp.

The difference between /var/tmp and /tmp from the point of view of *NIX
file-system hierarchy standard is twofold:

1) /var/tmp is expected to retain content across reboots
2) /var/tmp is expected to store more data than RAM allows

In practice /var/tmp must not be a RAM-based file system, unlike /tmp
which may, and in fact, is a RAM-based tmpfs on many contemporary
systems.

The primary motivation for this patch is QEMU, which uses /var/tmp as
scratch space for disk images running in snapshot mode, and as swap
memory for virtual RAM. QEMU tries very hard not to use /tmp for this
purpose and coercing it to use /tmp is both tricky and actively damages
usability of the project, as virtual machines may easily overflow main
memory when QEMU is asked not to persist changes to virtual storage.

This patch has a fatal flaw. Once landed, new snapd will immediately
grant access to /var/tmp from the point of view of the running snap
application process. Only newly constructed namespaces, however, will
get private /var/tmp. As such I'm not sure how to land it.

One terrible idea is to land the C changes without the Go changes, and
to land the Go changes at the next release. This would minimize, but not
eliminate, the chance that someone performs an in-place update.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>