o/state: add notice filter field for before timestamp
It can be useful to retrieve notices with last repeated timestamps
before a given time.

This introduces some complexity, as it may be the case that a notice is
recorded for the first time, then a filter is constructed with a before
timestamp after that notice's timestamp, and then before the notices
query occurs, the notice is re-recorded, thus giving it a last repeated
timestamp after the filter. The notice would then be omitted from the
list of notices returned by the query.

Additionally, WaitNotices no longer blocks indefinitely if it's
impossible for new notices to be recorded which satisfy the before
filter. This can be known if there are no notices matching the filter
when queried and the query occurs after the filter's before timestamp.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>