o/state: move last notice timestamp out of saved state
To support future notice creation which does not require the state lock,
add a dedicated lock for the last notice timestamp and add methods to
state to get the next monotonically increasing notice timestamp and to
update the last notice timestamp to a given timestamp, if it is later.

This way, other notice originators will be able to get a unique and
monotonically increasing notice timestamp without requiring state lock.
This is important for maintaining our API contracts, namely that any new
notice matching a given filter must have a timestamp newer than all
previous notices matching that filter.

Because the last notice timestamp may be updated without acquiring the
state lock, the updated last notice timestamp may not be saved to disk,
and so the timestamp on disk may become stale. For consistency, do not
store the last notice timestamp on disk, and instead re-compute it
during startup.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>