osutil: add AtomicWriteNoReplace flag to atomic file
It can be desirable to write a file only if it does not already exist.
Therefore, add an `AtomicWriteNoReplace` flag to indicate that this
should occur.

The actual implementation simply uses `os.Link()` in place of
`os.Rename()`, removing the file after adding the link in the former
case.

A more direct solution would be to use the `renameat2` syscall with the
`RENAME_NOREPLACE` flag set. However, this would require manually making
the syscall and handling any potential `EINTR`, and would rely on
filesystem support for the `RENAME_NOREPLACE` flag, which is not
universal.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>