Revert "s/apparmor: include files in apparmor kernel features (#15130)"
This reverts commit 666286f159884e56e1b9aaa0b8607cd6771bfa4c.

Including files from the apparmor features sysfs in the system key means
the system key changes more frequently when the kernel updates, which
causes problems for preseeding.

Furthermore, the only feature files which currently matter to snapd are
the `notify_versions` feature files, which only affect which protocol
version is used to communicate to the kernel -- they have no effect on
compiled profiles. So these do not need to be included in the reported
kernel features, and by extension the system key.

The checks which relied on the `notify_versions` feature files have
already been switched to check the filesystem directly as necessary,
thus the feature files can safely be removed from the probed features.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>