i/p/requestrules: add function to retrieve or set user session ID
Snapd needs a way to identify the systemd user session associated with a
given user, and detect when that user session terminates or restarts.

Systemd creates a tmpfs at `/run/user/<userID>` when the first session
for a given user starts. It persists and is shared by any concurrent
sessions for that same user, and when all sessions for that user
terminate, systemd deletes the tmpfs at `/run/user/<userID>` and
everything in it.

Therefore, snapd can use this tmpfs to identify the user session and
detect termination/restart by simply writing a random ID there.

Whenever snapd wants to use the current session ID for a given user, for
example to create a rule with lifespan "session" (coming soon) or detect
whether the user session associated with an existing rule has expired,
it can attempt to read the user session ID from `/run/user/<userID>`,
and if not found, generate and save a new one.

If the resulting session ID matches that of an existing rule, then we
know the session during which that rule was created is still active, and
if it differs, then we know that rule applies to a different user
session and should be treated as expired.

This commit adds a function to perform that load-or-store transaction,
returning either the existing user session ID or a newly-generated one,
which it has also written to disk. The function performs the read and
write while holding a lock, thus ensuring that there are no TOCTOU races
between multiple goroutines checking an absent user session ID file and
attempting to write a new one concurrently.

Because this function only operates within `/run/user`, which is a
tmpfs, there is no need for more robust atomic file operations which
would be necessary on a persistent filesystem. Snapd should be the only
writer to the user session ID file for any given user, and if it becomes
corrupted for some reason, it can be discarded and a new ID regenerated.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>