daemon: add action-based access checker
byActionAccess access checker decides which access checker
to use based on the JSON `action` field {"action": "<action>"}
to allow using different access checkers based on the target
action.

This is intended to be used by the /v2/system-volumes
endpoint but can also be used for similar action-based
endpoints if needed.

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>