s/apparmor: limit the extra information included in probed apparmor features and system key (#15624)
* s/apparmor: check notify version support files directly

When checking whether a particular notification protocol version is
supported, do not look through the list of probed AppArmor kernel
features. Instead check for the presence of notify version files
directly from the version support checks, and use the
`apparmor.MetadataTagsSupported()` function to check whether tagging is
supported, rather than performing a manual check.

This means that there will be no more reason to include the notify
version feature files in the probed AppArmor kernel features, and
furthermore, no reason to include any feature files at all, as no other
ones were ever checked.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* Revert "s/apparmor: include files in apparmor kernel features (#15130)"

This reverts commit 666286f159884e56e1b9aaa0b8607cd6771bfa4c.

Including files from the apparmor features sysfs in the system key means
the system key changes more frequently when the kernel updates, which
causes problems for preseeding.

Furthermore, the only feature files which currently matter to snapd are
the `notify_versions` feature files, which only affect which protocol
version is used to communicate to the kernel -- they have no effect on
compiled profiles. So these do not need to be included in the reported
kernel features, and by extension the system key.

The checks which relied on the `notify_versions` feature files have
already been switched to check the filesystem directly as necessary,
thus the feature files can safely be removed from the probed features.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor: only include specific entires from permstable32 in features

Rather than including every entry from
`/sys/kernel/security/apparmor/features/policy/permstable32` in the
probed AppArmor kernel features list, which is later included in the
system key, instead only include entries which are explicitly specified
to be included.

This reduces the opportunity for a system key mismatch due to changes in
extraneous permissions which do not necessarily affect profile
generation.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor: adjust naming and documentation

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>