o/state: save lastNoticeTimestamp on disk for backwards compatibility
In case of a revert/refresh to an older version of snapd which is not
aware that notices may live outside state, continue to write the
lastNoticeTimestamp to disk along with the rest of saved state.

This value can be relied upon to accurately reflect the last notice
timestamp of notices which are managed by snapd state, but it *cannot*
be relied upon to reflect the last notice timestamp of all notices from
backends other than state.

Since the last notice timestamp in state can be modified without holding
state lock, the value on disk can become stale unless state is
re-marshalled to disk for some other reason. For example, if the caller
of `NextNoticeTimestamp` holds state lock because they're adding a
notice to state, or some later operation on state requires the state
lock for writing.

During snapd startup, both snapd state and other notice backends should
update the last notice timestamp via the helper method in state or
abstractions over it in the upcoming notice manager.

Once other notice backends are defined, notices with types which should
no longer be managed by state should be migrated out of state and into
those other backends during startup. It will not be possible to support
the opposite (migrating notices from other backends back into state)
since older versions of snapd will not know about other notice backends.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>