o/state: add method to drain notices matching filter from state
As snapd is gaining the ability to have notice backends outside of
state, there is a need to ensure that existing notices can be migrated
from state to other notice backends when snapd is updated.

This commit adds a `DrainNotices` method which behaves much like
`Notices` except that it additionally removes the returned notices from
state. This method should only be called by the notice manager when a
new notice backend is being registered as a provider of a certain type
of notices, so the manager can transfer matching notices to the new
backend.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>