o/state: add notice helpers and before filter and move last notice timestamp out of saved state (#15601)
* o/state: add helpers for creating and re-recording notices

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: add notice filter field for before timestamp

It can be useful to retrieve notices with last repeated timestamps
before a given time.

This introduces some complexity, as it may be the case that a notice is
recorded for the first time, then a filter is constructed with a before
timestamp after that notice's timestamp, and then before the notices
query occurs, the notice is re-recorded, thus giving it a last repeated
timestamp after the filter. The notice would then be omitted from the
list of notices returned by the query.

Additionally, WaitNotices no longer blocks indefinitely if it's
impossible for new notices to be recorded which satisfy the before
filter. This can be known if there are no notices matching the filter
when queried and the query occurs after the filter's before timestamp.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: move last notice timestamp out of saved state

To support future notice creation which does not require the state lock,
add a dedicated lock for the last notice timestamp and add methods to
state to get the next monotonically increasing notice timestamp and to
update the last notice timestamp to a given timestamp, if it is later.

This way, other notice originators will be able to get a unique and
monotonically increasing notice timestamp without requiring state lock.
This is important for maintaining our API contracts, namely that any new
notice matching a given filter must have a timestamp newer than all
previous notices matching that filter.

Because the last notice timestamp may be updated without acquiring the
state lock, the updated last notice timestamp may not be saved to disk,
and so the timestamp on disk may become stale. For consistency, do not
store the last notice timestamp on disk, and instead re-compute it
during startup.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: simplify early return check for before filter in the past

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* fixup! o/state: add helpers for creating and re-recording notices

* o/state: expand test of before filter to return notices

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: save lastNoticeTimestamp on disk for backwards compatibility

In case of a revert/refresh to an older version of snapd which is not
aware that notices may live outside state, continue to write the
lastNoticeTimestamp to disk along with the rest of saved state.

This value can be relied upon to accurately reflect the last notice
timestamp of notices which are managed by snapd state, but it *cannot*
be relied upon to reflect the last notice timestamp of all notices from
backends other than state.

Since the last notice timestamp in state can be modified without holding
state lock, the value on disk can become stale unless state is
re-marshalled to disk for some other reason. For example, if the caller
of `NextNoticeTimestamp` holds state lock because they're adding a
notice to state, or some later operation on state requires the state
lock for writing.

During snapd startup, both snapd state and other notice backends should
update the last notice timestamp via the helper method in state or
abstractions over it in the upcoming notice manager.

Once other notice backends are defined, notices with types which should
no longer be managed by state should be migrated out of state and into
those other backends during startup. It will not be possible to support
the opposite (migrating notices from other backends back into state)
since older versions of snapd will not know about other notice backends.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: safely get last notice timestamp when marshalling state

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: clarify behavior of last notice timestamp when unmarshalling state

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: add method to drain notices matching filter from state

As snapd is gaining the ability to have notice backends outside of
state, there is a need to ensure that existing notices can be migrated
from state to other notice backends when snapd is updated.

This commit adds a `DrainNotices` method which behaves much like
`Notices` except that it additionally removes the returned notices from
state. This method should only be called by the notice manager when a
new notice backend is being registered as a provider of a certain type
of notices, so the manager can transfer matching notices to the new
backend.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: unexport method to get last notice timestamp

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>