i/prompting: add lifespan "session"
Replies and rules may have lifespan "session", which means they should
apply for the duration of the systemd user session associated with their
user ID. When that user session ends, any rules with lifespan "session"
which were associated with that session should be treated as expired.

As a result, the current user session ID is now passed to many prompting
helpers just as the current time has been, and is used to decide whether
a rule or ruel permission entry is expired, or to convert a permission
entry to a rule permission entry by assigning it a session ID. These
helpers handle much of the complexity around rules with lifespan
"session", and as long as the current user session for the corresponding
user is computed and passed into the helpers, much of the other code
within `interfaces/prompting/requestrules` needs minimal additional
changes.

Since the current user session ID is associated with the rule when a
rule with lifespan "session" is created, and snapd may restart while a
systemd user session remains active, the user session ID must be
marshalled to disk so it can be re-read after snapd restart.

As a side-effect, the user session ID for a given rule is also included
in the marshalled JSON in responses to API requests. This user session
ID is internal to snapd and meaningless to external applications.

Additionally, since a reply or rule creation could occur after the user
session ends (though this should never occur in practice, unless the
root user is replying/creating rules on behalf of a non-root user), an
error could occur when creating a rule with lifespan "session" due to
the session not being found. This error can be returned as a result of
API requests, so this commit adds a corresponding error kind and
handling for the error in the daemon.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>